Fix some erroneous definitions in the transitional xml fragment for 0-10.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@643472 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/broker/ConnectionHandler.h
@@ -53,7 +53,7 @@ class ConnectionHandler : public framing::FrameHandler
                      const std::string& mechanism, const std::string& response,
                      const std::string& locale); 
         void secureOk(const std::string& response); 
-        void tuneOk(uint16_t channelMax, uint32_t frameMax, uint16_t heartbeat); 
+        void tuneOk(uint16_t channelMax, uint16_t frameMax, uint16_t heartbeat); 
         void heartbeat() {}
         void open(const std::string& virtualHost,
                   const framing::Array& capabilities, bool insist);
